Add "findFile" function in fileOperands
Update introduces new "findFile" function, which helps you find one file (Or, if there are more files with the same name). Unlike the findFiles function, this does not take a keyword as a parameter, but rather takes a fixed name. This means that It is useful if you want to, say, find a file in a directory recursively called "main.txt", but there are other files in the directory such as "main12.txt" or "main200.txt", etc. So instead of keywords, it finds for the file with that specific name as passed in the parameter.